pyproject.toml
[project]
name = "algo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.42.1",
    "arrow>=1.3.0",
    "backtesting>=0.6.5",
    "loguru>=0.7.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]
